+++ src/spellbot/data.py
@@ -52,6 +52,7 @@ class Server(Base):
     expire = Column(Integer, nullable=False, server_default=text("30"))  # minutes
     links = Column(String(10), nullable=False, server_default=text("'public'"))
     power_enabled = Column(Boolean, nullable=False, server_default=true())
+    tags_enabled = Column(Boolean, nullable=False, server_default=true())
     create_voice = Column(Boolean, nullable=False, server_default=false())
     voice_category_xid = Column(BigInteger)
     games = relationship("Game", back_populates="server", uselist=True)